Moves threading helpers into a shared class, changes over things to use threading helper for combined yield/sleep
